Update mainline.sh
Returning "syntax error: unexpected end of file," rewriting to find the issue. I think I'm getting my "if then" statements wrong. I put in a second read so even if my variable check or write is coded wrong it should still move on to the second read function once the bug is patched. I'm going to test a few varieties on my own before updating the code in this branch.