Update mainline.sh
Added an ending for when the script is over. Added a new room. Discovered that Bash can't goto specific lines or call specific lines from memory, which was sort of the best way to write a complex text adventure in Batch with multiple rooms that could be revisited. All of the existence work-arounds that I found on StackExchange don't really work, either, I really needed the Goto function. I now realize that my text adventure will probably be incredibly linear, or better written in Assembly.